Increasing epsilon used for offest application
Older ffmpeg library requires that offest of each packet
is applied after seek, becaue it sends packets with
timestamps starting form seek to time, not staring from the beginning
of the clip.

Some times this causes problems in representation changing with
newer ffmpeg, because of the approximations of next segment of
new represenation differs from actual representation start.

This causes that after a representation switch offset is wrongly
added to timestamps.

To avoid it epsilon uses for offset need checkig was increased

Bug: https://cam.sprc.samsung.pl/browse/VDSMPNACL-1328
Change-Id: I372a3196fc4d36ed3404ef52033019f6040d1364